abstract trinket class to extend from
