+++ Content/Items/Trinkets/TestTrinket.cs
@@ -9,16 +9,13 @@ namespace IsaacTrinkets.Content.Items.Trinkets
 	// This is a basic item template.
 	// Please see tModLoader's ExampleMod for every other example:
 	// https://github.com/tModLoader/tModLoader/tree/stable/ExampleMod
-	public class TestTrinket : ModItem
+	public class TestTrinket : Trinket
 	{
 		// The Display Name and Tooltip of this item can be edited in the 'Localization/en-US_Mods.IsaacTrinkets.hjson' file.
 		public override void SetDefaults()
 		{
-			Item.width = 40;
-			Item.height = 40;
-			Item.accessory = true;
-			Item.value = Item.buyPrice(silver: 1);
-			Item.rare = ItemRarityID.Blue;
+			base.SetDefaults();
+			Item.defense = 10;
 		}
 
 		public override void AddRecipes()
@@ -27,24 +24,6 @@ namespace IsaacTrinkets.Content.Items.Trinkets
 			recipe.AddIngredient(ItemID.DirtBlock, 10);
 			recipe.AddTile(TileID.WorkBenches);
 			recipe.Register();
-		}
-
-		public override bool CanEquipAccessory(Player player, int slot, bool modded)
-		{
-			if (modded)
-			{
-				int moddedSlot = LoaderManager.Get<AccessorySlotLoader>().Get(ModContent.GetInstance<TrinketSlot>().Type, player).Type;
-
-				//Main.NewText("Modded slot: " + moddedSlot.ToString() + ", Selected slot: " + slot.ToString(), 255, 255, 255);
-
-				if (moddedSlot == slot)
-				{
-					return true;
-				}
-			}
-			
-
-			return false;
 		}
 	}
 }
